c8d: Implement push
This implements `docker push` under containerd image store.  When
pushing manifest lists that reference a content which is not present in
the local content store, it will attempt to perform the cross-repo mount
the content if possible.
Content that can't be repo mounted will be fetched from the repository
indicated by the `containerd.io/distribution.source` label.

Considering this scenario:

```bash
$ docker pull docker.io/library/busybox
```
This will download manifest list and only host platform-specific
manifest and blobs.

Now, tagging to a different repository (but still the same registry) and pushing:
```bash
$ docker tag docker.io/library/busybox docker.io/private-repo/mybusybox
$ docker push docker.io/private-repo/mybusybox
```

will still be able to push the whole manifest list, but will NOT
download any other platform-specific layer blobs. However, it will
fetch ALL platform manifests and images config because they can't be
cross repo mounted as they are manifests and not blobs. That shouldn't
be an issue though, because they are mostly small json files.

Doing the same, but with with a different registry:
```bash
$ docker tag docker.io/library/busybox some-random-repo.com/mybusybox
$ docker push some-random-repo.com/mybusybox
```

will not be able to utilize the cross-repo mount and WILL download all
blobs from the `docker.io/library/busybox` that are referenced by the
manifest list. This will effectively unshallow the image.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>