c8d/save: Include manifest.json for multiple platforms
Previously, `manifest.json` (which made the archive backwards compatible
with Docker image archive) was generated only if exported image was a
single platform manifest. Containerd selects only one manifest as the
`manifest.json`, so let's hint it to choose the default platform.

This makes the output archive compatible with both OCI format (which
allows to import all platforms) and Docker image which will import only
the platform it was exported from.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>