c8d/list: Make dangling images parseable by the cli
Without this, dangling images are not visible in the `docker image ls`
output.

This just restores behaviour introduced in "c8d/builder: store untagged
images" and accidentaly removed in recent PR which introduced changes
in image listing.

For upstreaming, should be squashed with:
fixup! Handle multi-platform images when listing them

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>